Allow the Job to re-create failed Pods
because when we try to upgrade, the Pod will be killed on reboot before
the job can reach "Complete" state. Thus, we want after reboot the Pod
to be re-created. The script is supposed to be idempotent. The second
time it should complete successfully allowing the node to be uncordoned
(because the Job will complete)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>