Use the operator image for the reboot Pod
and make sure the Pod can run a `reboot` and that it will be restarted
after reboot.

By using a separate Pod, we no longer require the user's workload to be
idempotent, a requirement posed by the fact that we doomed the Job to
fail, by running the reboot as part of it.

Now the reboot Pod, will come up in "Unknown" state after reboot,
Kubernetes will restart it and due to the sentinel annotations,
eventually the Pod will finish with exit code 0.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>